+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/web/controller/ProductController.java
@@ -1,7 +1,10 @@
 package gov.healthit.chpl.web.controller;
 
 
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,8 +22,11 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import gov.healthit.chpl.dao.EntityCreationException;
 import gov.healthit.chpl.dao.EntityRetrievalException;
 import gov.healthit.chpl.domain.Product;
+import gov.healthit.chpl.domain.ProductOwner;
 import gov.healthit.chpl.domain.UpdateProductsRequest;
+import gov.healthit.chpl.dto.DeveloperDTO;
 import gov.healthit.chpl.dto.ProductDTO;
+import gov.healthit.chpl.dto.ProductOwnerDTO;
 import gov.healthit.chpl.manager.ProductManager;
 import gov.healthit.chpl.manager.ProductVersionManager;
 import gov.healthit.chpl.web.controller.results.ProductResults;
@@ -105,13 +111,25 @@ public class ProductController {
 			}
 		} else {
 			if(productInfo.getProductIds().size() > 1) {
-				//if a product was send in, we need to do a "merge" of the new product and old products 
+				//if a product was sent in, we need to do a "merge" of the new product and old products 
 				//create a new product with the rest of the passed in information
 				ProductDTO newProduct = new ProductDTO();
 				newProduct.setName(productInfo.getProduct().getName());
 				newProduct.setReportFileLocation(productInfo.getProduct().getReportFileLocation());
 				if(productInfo.newDeveloperId() != null) {
 					newProduct.setDeveloperId(productInfo.newDeveloperId());
+				}
+				//new product could be created with ownership history
+				if(productInfo.getProduct().getOwnerHistory() != null) {
+					for(ProductOwner prevOwner : productInfo.getProduct().getOwnerHistory()) {
+						ProductOwnerDTO prevOwnerDTO = new ProductOwnerDTO();
+						prevOwnerDTO.setId(prevOwner.getId());
+						DeveloperDTO dev = new DeveloperDTO();
+						dev.setId(prevOwner.getDeveloper().getDeveloperId());
+						prevOwnerDTO.setDeveloper(dev);
+						prevOwnerDTO.setTransferDate(prevOwner.getTransferDate());
+						newProduct.getOwnerHistory().add(prevOwnerDTO);
+					}
 				}
 				result = productManager.merge(productInfo.getProductIds(), newProduct);
 				
@@ -125,6 +143,19 @@ public class ProductController {
 				if(productInfo.newDeveloperId() != null) {
 					toUpdate.setDeveloperId(productInfo.newDeveloperId());
 				}
+				//product could have updated ownership history
+				if(productInfo.getProduct().getOwnerHistory() != null) {
+					for(ProductOwner prevOwner : productInfo.getProduct().getOwnerHistory()) {
+						ProductOwnerDTO prevOwnerDTO = new ProductOwnerDTO();
+						prevOwnerDTO.setId(prevOwner.getId());
+						prevOwnerDTO.setProductId(toUpdate.getId());
+						DeveloperDTO dev = new DeveloperDTO();
+						dev.setId(prevOwner.getDeveloper().getDeveloperId());
+						prevOwnerDTO.setDeveloper(dev);
+						prevOwnerDTO.setTransferDate(prevOwner.getTransferDate());
+						toUpdate.getOwnerHistory().add(prevOwnerDTO);
+					}
+				}
 				result = productManager.update(toUpdate);
 			}	
 		}
@@ -132,6 +163,9 @@ public class ProductController {
 		if(result == null) {
 			throw new EntityCreationException("There was an error inserting or updating the product information.");
 		}
-		return new Product(result);
+		
+		//get the updated product since all transactions should be complete by this point
+		ProductDTO updatedProduct = productManager.getById(result.getId());
+		return new Product(updatedProduct);
 	}
 }