+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/dao/impl/ProductDaoTest.java
@@ -8,6 +8,7 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.test.annotation.Rollback;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestExecutionListeners;
@@ -25,6 +26,7 @@ import gov.healthit.chpl.auth.user.JWTAuthenticatedUser;
 import gov.healthit.chpl.dao.EntityRetrievalException;
 import gov.healthit.chpl.dao.ProductDAO;
 import gov.healthit.chpl.dto.ProductDTO;
+import gov.healthit.chpl.dto.ProductOwnerDTO;
 import gov.healthit.chpl.entity.ProductEntity;
 import junit.framework.TestCase;
 
@@ -40,17 +42,20 @@ public class ProductDaoTest extends TestCase {
 	@Autowired
 	private ProductDAO productDao;
 
-	private static JWTAuthenticatedUser authUser;
-
+	private static JWTAuthenticatedUser adminUser;
+	
 	@BeforeClass
 	public static void setUpClass() throws Exception {
-		authUser = new JWTAuthenticatedUser();
-		authUser.setFirstName("Admin");
-		authUser.setId(-2L);
-		authUser.getPermissions().add(new GrantedPermission("ROLE_ADMIN"));
+		adminUser = new JWTAuthenticatedUser();
+		adminUser.setFirstName("Administrator");
+		adminUser.setId(-2L);
+		adminUser.setLastName("Administrator");
+		adminUser.setSubjectName("admin");
+		adminUser.getPermissions().add(new GrantedPermission("ROLE_ADMIN"));
 	}
 
 	@Test
+	@Transactional(readOnly = true)
 	public void getAllProducts() {
 		List<ProductDTO> results = productDao.findAll();
 		assertNotNull(results);
@@ -58,6 +63,7 @@ public class ProductDaoTest extends TestCase {
 	}
 
 	@Test
+	@Transactional(readOnly = true)
 	public void getProductById() {
 		Long productId = -1L;
 		ProductDTO product = null;
@@ -71,6 +77,7 @@ public class ProductDaoTest extends TestCase {
 	}
 	
 	@Test
+	@Transactional(readOnly = true)
 	public void getProductByDeveloper() {
 		Long developerId = -1L;
 		List<ProductDTO> products = null;
@@ -80,6 +87,7 @@ public class ProductDaoTest extends TestCase {
 	}
 	
 	@Test
+	@Transactional(readOnly = true)
 	public void getProductByDevelopers() {
 		List<Long> developerIds = new ArrayList<Long>();
 		developerIds.add(-1L);
@@ -91,11 +99,15 @@ public class ProductDaoTest extends TestCase {
 	}
 	
 	@Test
+	@Transactional
+	@Rollback(true)
 	public void updateProduct() throws EntityRetrievalException {
+		SecurityContextHolder.getContext().setAuthentication(adminUser);
+
 		ProductDTO product = productDao.getById(-1L);
 		product.setDeveloperId(-2L);
 		
-		ProductEntity result = null;
+		ProductDTO result = null;
 		try {
 			result = productDao.update(product);
 		} catch(Exception ex) {
@@ -111,6 +123,7 @@ public class ProductDaoTest extends TestCase {
 			fail("could not find product!");
 			System.out.println(ex.getStackTrace());
 		}
+		SecurityContextHolder.getContext().setAuthentication(null);
 	}
 	
 	/** Description: Tests that the productDAO can create a new product based on the productDTO
@@ -124,6 +137,8 @@ public class ProductDaoTest extends TestCase {
 	@Rollback(true)
 	@Test
 	public void createProduct() throws EntityRetrievalException {
+		SecurityContextHolder.getContext().setAuthentication(adminUser);
+
 		ProductDTO product = new ProductDTO();
 		product.setCreationDate(new Date());
 		product.setDeleted(false);
@@ -144,5 +159,7 @@ public class ProductDaoTest extends TestCase {
 		assertNotNull(result.getId());
 		assertTrue(result.getId() > 0L);
 		assertNotNull(productDao.getById(result.getId()));
+	
+		SecurityContextHolder.getContext().setAuthentication(null);
 	}
 }