+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/manager/impl/DeveloperManagerTest.java
@@ -1,5 +1,8 @@
 package gov.healthit.chpl.manager.impl;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -12,6 +15,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.support.DependencyInjectionTestExecutionListener;
 import org.springframework.test.context.support.DirtiesContextTestExecutionListener;
 import org.springframework.test.context.transaction.TransactionalTestExecutionListener;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.github.springtestdbunit.DbUnitTestExecutionListener;
@@ -25,8 +29,11 @@ import gov.healthit.chpl.dao.EntityRetrievalException;
 import gov.healthit.chpl.dto.DeveloperACBMapDTO;
 import gov.healthit.chpl.dto.DeveloperDTO;
 import gov.healthit.chpl.dto.DeveloperStatusDTO;
+import gov.healthit.chpl.dto.ProductDTO;
+import gov.healthit.chpl.dto.ProductOwnerDTO;
 import gov.healthit.chpl.entity.DeveloperStatusType;
 import gov.healthit.chpl.manager.DeveloperManager;
+import gov.healthit.chpl.manager.ProductManager;
 import junit.framework.TestCase;
 
 @RunWith(SpringJUnit4ClassRunner.class)
@@ -39,6 +46,7 @@ import junit.framework.TestCase;
 public class DeveloperManagerTest extends TestCase {
 	
 	@Autowired private DeveloperManager developerManager;
+	@Autowired private ProductManager productManager;
 	@Autowired private DeveloperStatusDAO devStatusDao;
 	
 	private static JWTAuthenticatedUser adminUser;
@@ -86,6 +94,32 @@ public class DeveloperManagerTest extends TestCase {
 	}
 	
 	@Test
+	@Transactional
+	public void testGetAllDevelopersIncludingDeletedAsAdmin() throws EntityRetrievalException {
+		SecurityContextHolder.getContext().setAuthentication(adminUser);
+		List<DeveloperDTO> developers = developerManager.getAllIncludingDeleted();
+		assertNotNull(developers);
+		assertEquals(9, developers.size());
+		SecurityContextHolder.getContext().setAuthentication(null);
+	}
+	
+	@Test
+	@Transactional
+	public void testGetAllDevelopersIncludingDeletedUncredentialed() throws EntityRetrievalException {
+		SecurityContextHolder.getContext().setAuthentication(null);
+		List<DeveloperDTO> developers = null;
+		boolean failed = false;
+		try {
+			developers = developerManager.getAllIncludingDeleted();
+		} catch(Exception ex) {
+			//should fail
+			failed = true;
+		}
+		assertNull(developers);
+		assertTrue(failed);
+	}
+	
+	@Test
 	@Rollback
 	public void testDeveloperStatusChangeAllowedByAdmin() 
 			throws EntityRetrievalException, JsonProcessingException {
@@ -105,6 +139,58 @@ public class DeveloperManagerTest extends TestCase {
 		assertFalse(failed);
 		assertNotNull(developer.getStatus());
 		assertEquals(DeveloperStatusType.SuspendedByOnc.toString(), developer.getStatus().getStatusName());
+		SecurityContextHolder.getContext().setAuthentication(null);
+	}
+	
+	@Test
+	@Rollback(true)
+	public void testMergeDeveloper_productOwnershipHistoryAdded() {
+		SecurityContextHolder.getContext().setAuthentication(adminUser);
+
+		List<Long> idsToMerge = new ArrayList<Long>();
+		idsToMerge.add(-1L);
+		idsToMerge.add(-4L);
+		
+		DeveloperDTO toCreate = new DeveloperDTO();
+		toCreate.setName("dev name");
+		
+		DeveloperDTO merged = null;
+		try {
+			merged = developerManager.merge(idsToMerge, toCreate);
+		} catch(EntityCreationException | JsonProcessingException | EntityRetrievalException ex) {
+			fail(ex.getMessage());
+		}
+		
+		assertNotNull(merged);
+		assertNotNull(merged.getId());
+		
+		try {
+			ProductDTO affectedProduct = productManager.getById(-3L);
+			assertNotNull(affectedProduct);
+			assertNotNull(affectedProduct.getOwnerHistory());
+			assertEquals(1, affectedProduct.getOwnerHistory().size());
+			assertEquals(-1, affectedProduct.getOwnerHistory().get(0).getDeveloper().getId().longValue());
+		} catch(EntityRetrievalException ex) {
+			fail(ex.getMessage());
+		}
+		
+		try {
+			ProductDTO affectedProduct = productManager.getById(-1L);
+			assertNotNull(affectedProduct);
+			assertNotNull(affectedProduct.getOwnerHistory());
+			assertEquals(2, affectedProduct.getOwnerHistory().size());
+			int expectedDevsCount = 0;
+			for(ProductOwnerDTO owner : affectedProduct.getOwnerHistory()) {
+				if(owner.getDeveloper() != null && owner.getDeveloper().getId().longValue() == -1) {
+					expectedDevsCount++;
+				} else if(owner.getDeveloper() != null && owner.getDeveloper().getId().longValue() == -2) {
+					expectedDevsCount++;
+				}
+			}
+			assertEquals(2, expectedDevsCount);
+		} catch(EntityRetrievalException ex) {
+			fail(ex.getMessage());
+		}
 		SecurityContextHolder.getContext().setAuthentication(null);
 	}