+++ RELEASE_NOTES.md
@@ -1,5 +1,18 @@
 # Release Notes
 
+## Version 6.0.0
+_15 November 2016_
+
+### Features Added
+* Add ability to retrieve and edit product ownership history (which developers a product has previously been owned by). BREAKS BACKWARDS COMPATIBILITY when retrieving certified product details
+* Add "showDeleted" parameter to /developers call. Defaults to false.
+
+### Bugs Fixed
+* Make it impossible to confirm duplicate pending certified products
+* Ability to edit a certified product without security authorization
+
+--- 
+
 ## Version 5.2.0
 _21 October 2016_